requirements.txt
discord
requests
datetime
asyncio
colorama
thread6
requests_oauthlib
flask
pillow
psutil